+++ CHANGELOG.rst
@@ -1,6 +1,22 @@
 lib_xua change log
 ==================
 
+UNRELEASED
+----------
+
+  * ADDED:    Optional defines to provide insertion of user endpoints, and
+    handling code for them allowing composite USB Audio devices to be built,
+    such as CDC. These are provided by optional include headers 
+    ``xua_user_descriptors_incl.h``, ``xua_user_endpoint_init.h``,
+    ``xua_user_descriptors_content.h``, ``xua_user_endpoint0_decl.h``,
+    ``xua_user_descriptors_decl.h`` and ``xua_user_endpoint0_handler.h`` as
+    well as the defines ``XUA_USER_IN_ENDPOINTS``, ``XUA_USER_OUT_ENDPOINTS``,
+    and ``XUA_USER_INTERFACES``.
+
+  * ADDED:     Optional ``XUA_USER_FUNCTION_CALL_PRE_BUFFER`` define to allow
+    running of user code before XUA buffer starts in the ``main()`` function.
+    This may be used to initialise global channel ends or interfaces.
+
 5.2.0
 -----